// stb-backend/src/common/enums/employee-status.enum.ts
export enum EmployeeStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SUSPENDED = 'SUSPENDED',
  PENDING_ACTIVATION = 'PENDING_ACTIVATION',
  LOCKED = 'LOCKED',
  ON_LEAVE = 'ON_LEAVE',
  TERMINATED = 'TERMINATED',
  PROBATION = 'PROBATION',
  BLOCKED = 'BLOCKED',
  DEACTIVATED = 'DEACTIVATED',
}